Fix SQL bom properly
This fixes the root cause for https://github.com/micronaut-projects/micronaut-core/issues/7068
which is that the BOM we generated contained a constraint without version.

Constraints without versions do not make sense: the role of a constraint
_is_ to influence dependency resolution by providing at least a version.
This should have been caught by Gradle and should in any case not trigger
a parsing bug (see https://github.com/gradle/gradle/issues/20182) but
that's how it is.