Require Java 11 for micronaut-jooq and patch update jooq
In 4.5.0 we upgraded Jooq to get R2DBC support.

As Jooq 3.15.x requires Java 11, the Jooq module was marked as Java 11 only (in v4.5.x)

This restriction was removed in 4.6.x, however jOOQ still required Java 11.

Normally, we would downgrade jOOQ to maintain Java 8 support, however R2DBC support only exists in 3.15+ so we cannot do this.

This PR:

- re-introduces the java 11 compile target for the jOOQ module
- adds a note to the documentation to say that the jooq module requires Java 11+
- Updates jOOQ to the latest patch release of 3.15 (at time of writing)

Closes #821