Check if Metrics are enabled prior to getting the MeterRegistry bean
We had an issue of a circular loop when trying to load a DataSource by name from the BeanContext.

Disabling the metrics made no difference, as the Bean was retrieved prior to checking if it was enabled.

This change checks that metrics are enabled prior to attempting to get the metrics bean.